+++ go/genesis-tools/main.go
@@ -1,0 +1,10 @@
+// Stats implementation.
+package main
+
+import (
+	"github.com/oasislabs/the-quest-entities/go/genesis-tools/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
