+++ go/genesis-tools/staking.json
@@ -1,0 +1,1 @@
+{"params":{"thresholds":{"0":"100000000000000000000","1":"100000000000000000000"},"debonding_interval":1,"commission_schedule_rules":{},"gas_costs":{"add_escrow":1000,"burn":1000,"reclaim_escrow":1000,"transfer":1000},"min_delegation":"10000000000000000000"},"total_supply":"10000000000000000000000000000","common_pool":"7999999000000000000000000000","ledger":{"ABC2K2tpD5SiQGAHSXUMnJpvdouFnKAHayK4zY118uo=":{"general":{"balance":"2000000000000000000000000000","nonce":0},"escrow":{"active":{"balance":"0","total_shares":"0"},"debonding":{"balance":"0","total_shares":"0"},"commission_schedule":{"rates":null,"bounds":null}}},"Pr32K2tpD5SiQGAHSXUMnJpvdouFnKAHayK4zY118uo=":{"general":{"balance":"0","nonce":0},"escrow":{"active":{"balance":"500000000000000000000","total_shares":"500000000000000000000"},"debonding":{"balance":"0","total_shares":"0"},"commission_schedule":{"rates":null,"bounds":null}}},"uOHHb0wIBW0cumq8nzffWkOJFUx4PR++rMHp5HR5Aio=":{"general":{"balance":"0","nonce":0},"escrow":{"active":{"balance":"500000000000000000000","total_shares":"500000000000000000000"},"debonding":{"balance":"0","total_shares":"0"},"commission_schedule":{"rates":null,"bounds":null}}}},"delegations":{"Pr32K2tpD5SiQGAHSXUMnJpvdouFnKAHayK4zY118uo=":{"Pr32K2tpD5SiQGAHSXUMnJpvdouFnKAHayK4zY118uo=":{"shares":"500000000000000000000"}},"uOHHb0wIBW0cumq8nzffWkOJFUx4PR++rMHp5HR5Aio=":{"uOHHb0wIBW0cumq8nzffWkOJFUx4PR++rMHp5HR5Aio=":{"shares":"500000000000000000000"}}}}
