--- go/staking-ledger/main.go
@@ -1,10 +1,0 @@
-// Stats implementation.
-package main
-
-import (
-	"github.com/oasislabs/the-quest-entities/go/staking-ledger/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
